(svn r26596) [1.4] -Backport from trunk: - Fix: Save/load issues on big endian machines (r26593, r26590, r26589) - Fix: Consider multiheaded trains in station refits [FS#5995] (r26586) - Fix: Game script could be changed in game by double clicking [FS#5974] (r26583)

+++ src/saveload/linkgraph_sl.cpp
@@ -21,7 +21,7 @@ typedef LinkGraph::BaseEdge Edge;
 
 const SettingDesc *GetSettingDescription(uint index);
 
-static uint _num_nodes;
+static uint16 _num_nodes;
 
 /**
  * Get a SaveLoad array for a link graph.
